Update README.md
# Current_Affairs

The application is set to save people with current affairs. News that surrounds us and from abroad. The application allows a user to signup and the administrator is rresponsible in updating the news, deleting news and comments from the user on  a basis that the news does not follow rights of the community.